chore: move constraint related type to separate file
