+++ src/validator/constraints/other/compound.ts
@@ -1,5 +1,6 @@
-import type { ConstraintContext, FieldRules } from '../../types';
+import type { FieldRules } from '../../types';
 import { Violation } from '@/validator/violation';
+import { ConstraintContext } from '@/validator/constraint';
 
 export function compound(rules: FieldRules) {
   return function compoundConstraint(value: unknown, context: ConstraintContext): Violation[] {
