+++ src/validator/constraints/string/email.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
-import type { ConstraintContext } from '../../types';
 import { email } from './email';
+import { ConstraintContext } from '@/validator/constraint';
 
 describe('email', () => {
   it('rejects invalid email by default', () => {
