+++ src/validator/factory/create-validator.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest';
 import { UnknownConstraintError } from '../errors';
-import type { ConstraintContext } from '../types';
 import { createValidator } from './create-validator';
 import { Violation } from '@/validator/violation';
+import { ConstraintContext } from '@/validator/constraint';
 
 describe('Validator', () => {
   describe('Registry', () => {
