+++ src/validator/types.ts
@@ -1,17 +1,7 @@
 import { Violation } from './violation';
 import { Payload } from './payload';
-import { Options } from '@/validator/options';
-
-export type ConstraintOptions = { groups?: string[] } & Record<string, unknown>;
-export type ConstraintContext = {
-  path: string;
-  root: unknown;
-  value: unknown;
-  constraint: string;
-  options: ConstraintOptions;
-  runNestedRules: (value: unknown, rules: FieldRules, path: string) => Violation[];
-};
-export type ConstraintValidator = (value: unknown, context: ConstraintContext) => Violation[];
+import { Options } from './options';
+import { ConstraintOptions } from './constraint';
 
 export type FieldRuleEntry = string | Record<string, ConstraintOptions>;
